rand(a:b) : usually faster, except for immediate a and b
We introduce a third sampler for big a or b, but short b-a.
We can then use a regular range sampler for bitstypes.
The goal was also for rand to not depend on the "reduce" stuff,
i.e. to not store a short integer in an array, which was needed
before in order to call `MPN.cmp`.
But this causes the "fast path" for small a and b to be
roughly 2x slower.