native `add!` for mixed `XInt` and `BitInteger`
We revert the previous work which made this possible for `SLimbMax`,
which was at the cost of slightly increased complexity, as we had
to handle the case `abs(x) >= limb1min`. So we get back the usual
assumptions, and instead wrap a bit integer as a "XInt-like" object
(`Static`), so that the previous algorithms can work on them.
This implies:
1. re-implement in Julia some MPN routines, which require heap
   allocated limbs in GMP;
2. don't swap values (e.g. when we need `x >= y`, but we have `x < y`);
   instead, have helper functions which we call with a different
   order of arguments; this is needed because `x`, `y` might have
   different types (e.g. `XInt` and `Static`), and we don't want
   to create type-unstability by swapping them.